mpris: Read track length from Word64 too
According to the MPRIS v2 spec, the length of a track "must be given in
microseconds, and be represented as a signed 64-bit integer". [1]

But Spotify does not follow the spec and represents it as an unsigned
64-bit integer:

```
$ dbus-send --session --print-reply --reply-timeout=150 --dest=org.mpris.MediaPlayer2.spotify /org/mpris/MediaPlayer2 org.freedesktop.DBus.Properties.Get string:org.mpris.MediaPlayer2.Player string:Metadata

method return time=1743433787.301824 sender=:1.142 -> destination=:1.178 serial=1071 reply_serial=2
   variant       array [
         dict entry(
            string "mpris:length"
            variant                uint64 152000000
         )
...
```

This always made the `length` template argument end up empty, but
allowing a Word64 for this attribute fixes this problem.

[1]: https://specifications.freedesktop.org/mpris-spec/latest/Track_List_Interface.html#Mapping:Metadata_Map