feat: add new PacmanUpdates constructor for ArchUpdates Plugin
the new PacmanUpdates constructor behaves similar to the ArchUpdates
constructor while additionally allowing to pass in a custom error
message for unknown pacman  failures. The default error message of
`pacman: Unknown cause of failure.` of the ArchUpdates ctor is too
long for my taste.

The ArchUpdates ctor was modified to delegate to the new PacmanUpdates
plugin while providing the default error message keeping existing behavior
intact.

The new constructors  name better represents the Plugin's compatibility
with all pacman-based distributions, not just Arch.

The docs have been updated to reflect the existence of the new plugin
and to highlight the similarities between the ArchUpdates and
PacmanUpdates constructor.