docs: make description of PacmanUpdates plugin more clear
There is now need for a `?` to be in the `Many` string at all. It's just
a possibility for the user. I personally am using a different NerdFont
Icon for One, Zero, Many and Error and don't include the `?` in the
string. This doc change reflects the possibilty to provide a string with
or without a `?` in the docs.